Update how to handle reference node
When updating call node reference node, it was
not being checked if a loop contained a call
node. So, each node that was calling a function
did not have its reference node updated.

Tests were also updated and get_loop_wcec()
function was changed to get_refnode_rwcec(). In
this way, reference node is more generic if it is
a reference to an entry node or a loop.