Remove else-if statements
All reference for else-if statements were
removed. Now, if there is a else-if statement,
then it should be an else statement following by
an if statement. Such as:

from:
    else if (...) {...}
to:
    else {
        if (...) {...}
    }

In this way, DVFS code can be inserted before if
statement since there is another line which
differentiate the else and if statements. So, in
the previous way when else and if were together,
no DVFS code could be inserted before the
condition of if.